moved commander options to new file, work more on typescript componnets, fixed some component template lines, comments
